requirements.txt
flask
fpdf
requests
pillow
gunicorn
